i41: Move screens and components to src/library
