+++ src/screens/mapbox/index.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import { StyleSheet, View, PermissionsAndroid, Platform } from 'react-native';
 import StaticSafeAreaInsets from 'react-native-static-safe-area-insets';
-import config from '../../config';
+import config from '../../../config';
 import MapboxGL from '@react-native-mapbox-gl/maps';
-import userMarker from '../assets/img/user_marker.png';
+import userMarker from '../../../frontend/assets/img/user_marker.png';
 import Geolocation from 'react-native-geolocation-service';
-import {bindMethods} from '../component-ops';
-import FAB from '../components/fab';
+import {bindMethods} from '../../../frontend/component-ops';
+import FAB from '../../library/components/fab';
 import _ from 'lodash';
 import CompassHeading from 'react-native-compass-heading';
